requirements.txt
colorlog
structlog
python-dotenv
Flask
flask-cors
flask-marshmallow
Flask-Migrate
flask-restful
Flask-SQLAlchemy
marshmallow-sqlalchemy
psycopg2-binary
PyJWT
gunicorn
requests
